refactor: better package/folder naming and move in process server into back end
