+++ acceptance/screenplay/screenplay.go
@@ -13,7 +13,7 @@ type Question func(Abilities) (interface{}, error)
 
 type Abilities struct {
 	Name      string
-	App       driver.AcceptanceTestDriver
+	App       driver.TestDriver
 	LastError error
 }
 
@@ -32,7 +32,7 @@ type Actor struct {
 	abilities Abilities
 }
 
-func NewActor(name string, app driver.AcceptanceTestDriver) *Actor {
+func NewActor(name string, app driver.TestDriver) *Actor {
 	ret := &Actor{
 		abilities: Abilities{
 			Name: name,
