+++ acceptance/suite_test.go
@@ -11,7 +11,7 @@ import (
 
 type suite struct {
 	actors map[string]*screenplay.Actor
-	driver driver.AcceptanceTestDriver
+	driver driver.TestDriver
 }
 
 func (s *suite) Actor(name string) *screenplay.Actor {
@@ -21,7 +21,7 @@ func (s *suite) Actor(name string) *screenplay.Actor {
 	return s.actors[name]
 }
 
-func RunSuite(t *testing.T, driver driver.AcceptanceTestDriver, featurePaths []string) {
+func RunSuite(t *testing.T, driver driver.TestDriver, featurePaths []string) {
 	suite := godog.TestSuite{
 		ScenarioInitializer: func(ctx *godog.ScenarioContext) {
 			s := &suite{
