+++ back-end/pkg/entities/entities.go
@@ -1,3 +1,4 @@
+// Entities package is exported so that it can be reused in acceptance tests
 package entities
 
 type Project struct {
